requirements.txt
numpy
matplotlib
pillow
zarr
numcodecs
tqdm
torch
pyyaml
l5kit    # optional (only if you want to use real Lyft dataset)
